+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 - fix transformation in Julia: nthRoot, exponentiale, zero dynamic, ode_ priority, factorial
 - add bind() method to `DSwitcher`
-- export `DSwitcher` to formats: SLV, DBSolve, Matlab, Simbiology
+- export `DSwitcher` to formats: SLV, DBSolve, Matlab, Simbiology, SBML
 - multispace Julia export
+- export `CSwitcher` to formats: SLV, DBSolve, SBML
 
 ## 0.5.11